Add min heights, stylize header, stylize illustrations/footer
Stylize navbar to reflect actual text instead of placeholders, and add
a hover pseudoclass behavior to change color of navbar text when
hovering over the items there.

Add min heights to sections to avoid distortion when shrinking browser.

Stylize header text to include appropriate font size, weight, and color.
Stylize the button in header to proper font style and border properties.

Adjust the margins of the illustrations container and the margins of
the items within. Add a border to the images, a background color and a
drop shadow. Replace placeholder subtext with actual text, then adjust
the margins and positioning of that text. Install a hover behavior on
the items in the illustration items that slightly scales up the image.

Add a matching background color to the quote section to give separation
of section. Stylize text using a variety of font-sizes, font styles,
and weights. Adapt to work with rudimentary 'mobile-friendy' code
detailed below.

Re-parent banner container section to a new section called
'banner-main' to allow for alignment of banner. The previous 'banner'
section into a div and have it be within the new 'banner' main section.
Place banner text and banner button into flex container and stylize the
text with font-size and font weight. Then stylize the button similarly
to the previous button, but with a white border color and text color,
as well as with a hover behavior to lighten the background color of the
button.

Set footer to a specific size to avoid it being the same size asa the
text within it. Set the container in whicn the text resides to the same
size as the container, then center and aligh the text within the
container. Stylize the font size of the footer.

Add a media query that changes the behavior of the various flex
directions within the sections to column. Set the illustrations section
to display: block to quickly make the images align vertically stacked.
Set the media query to activate once the screen size falls under 768 px,
and these sections will align vertically to make the page friendy to
portrait screens such as mobile.